refactor report
