+++ pkg/collector/corechecks/snmp/common/testing.go
@@ -1,8 +1,8 @@
-package snmp
+package common
 
 import "time"
 
-var mockTimeNow = func() time.Time {
+var MockTimeNow = func() time.Time {
 	layout := "2006-01-02 15:04:05"
 	str := "2000-01-01 00:00:00"
 	t, _ := time.Parse(layout, str)
