+++ pkg/collector/corechecks/snmp/report/report_bandwidth_usage.go
@@ -1,4 +1,4 @@
-package snmp
+package report
 
 import (
 	"fmt"
